requirements.txt
nicegui
openai
requests
python-dotenv
sqlalchemy
bcrypt
